Assignment_v04
Underway with the implementation of the millis() function in attempting
to animate the text on screen. A new file "blinks was added" to simply
add the large portion of code that allows the underscore to blink off
and on, which will be reduced at a later point. Currently having trouble
with the underscore not conforming to the fill colour it is set.