Assignment_v06
The initialising string that is printed i=in the centre of the screen
has been fixed so that the spacing between the letters is now correct.
It also animates at a good speed now. The final part will be to let the
underscore (cursor) blink a couple of times before the screen will go
black again and the UI will come into focus.