Assignment_v12
Previously work was being made towards a shooting star function, however
this has been removed for a better aesthetic. Now work is being done for
a system in which when a star is being hovered over a box will appear
with a description of that star. The current code is not functioning
100% and still has bugs.