Assignment_v13
Progress is being made towards providing information on each star. A new
class has been made to create many different boxes for now and the rest
of the project. The previous bugs with drawing the circles has been
worked out and lines are extending appropriately depending on their
position in the scene.