Assignment_v25
Moved some of the code in draw to a seperate function to clear up and
make more obvious each stage of the program. Improvements have meen made
to the temperature guage in that it now is coloured to represent the
temperature. An actual value will be placed within the notches
displaying the temperature.