Assignment_v30
A start has been made towards functions to expand the buttons into menus
where certain funtions can take place. The draw button function in the
Button class was also added inorder to print the buttons in the ship
screen. The stroke value throughout the functions relating to the ship
screen have been upped to 100, to make it slightly brighter for the
user.